Add support for kernel and initrd image

+++ board/mangopi/common/script/post-build.sh
@@ -6,7 +6,12 @@ echo "nameserver 4.2.2.4" >> $1/etc/resolv.conf
 
 echo "none		/sys/kernel/debug		debugfs	defaults	0	0" >> $1/etc/fstab
 
-cp $BOARD_DIR/kernel.its $BINARIES_DIR/kernel.its
+if test -e $BINARIES_DIR/rootfs.cpio.xz ; then
+	cp $BOARD_DIR/kernel_initrd.its $BINARIES_DIR/kernel.its
+else 
+	cp $BOARD_DIR/kernel.its $BINARIES_DIR/kernel.its
+fi
+
 (cd $BINARIES_DIR; $HOST_DIR/bin/mkimage -f kernel.its kernel.itb; rm kernel.its)
 
 echo "$BRD end"
